pyproject.toml
[tool.poetry]
authors = ["m09 <142691+m09@users.noreply.github.com>"]
description = ""
name = "tp-bibliotheque"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6"
sqlalchemy = "^2"

[tool.poetry.group.dev.dependencies]
black = "^24"
flake8 = "^7"
flake8-bugbear = "^24"
flake8-docstrings = "^1"
isort = "^5"
mypy = "^1"
pylint = "^3"
pytest = "^8"
types-pyyaml = "^6"

[tool.poetry.scripts]
bibli = "tp_bibliotheque.cli:main"

[tool.isort]
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
